Carve `TestCorpusFixtures` out of the monolith as one unit (wala/ML#635).
The 17 whole-project corpus tests (vendored NLPGNN, gpt-2,
MusicTransformer, BiLSTM, TextCNN), the `runNlpgnnFullGeneration`
helper, and both fixture arrays move together: they are the suite's
long pole (wala/ML#755) and their loop-carried expectations are
order-sensitive (wala/ML#753), so they stay one class.
`DiagnosticLoggingVolumeTest` and `WorklistOrderInvarianceTest`
retarget to the new class. The monolith keeps 397, preserving the
1,040 total.
